Switch out termion for crossterm
The way termion stores terminal state is flawed. For example, entering
raw mode, the previous state is stored locally in a struct, whereas
crossterm stores it in a lazy static.

There are several problems with the termion way of doing it:

1. Raw mode can be 'entered' multiple times

Using 2 different references to stdout, it is possible to 'enter' raw
mode using both of them. The first one's previous state will be cooked
mode. The second will be... raw mode. Subsequently calling
'suspend_raw_mode' on the second will do nothing.

2. It not possible to properly recover from panics

The struct which allows resetting the terminal state cannot be shared,
so I can't figure out a way to use it to restore state both in normal
operation (when dropped) _and_ in a panic hook handler.

Crossterm's global state fixes both of these problems, and is more
representative of the fact that we _do_ have some essentially global
external state.