Implement page_up/page_down
This was a bit tricky, and the solution feels a bit hacky.

`tui` doesn't give easy access to sizes of different parts of the
layout, because the layout isn't supposed to be stateful, it's figured
out on demand when rendering.

However, implementing page_up/page_down requires knowing the list
height... awkward.